Move validation to delivery
